Add deleteComment function
